refactor: rename llm.py to llm_model.py and migrate LLM classes for improved organization
